remove editor file
